Updated step() to take a N_max_steps, interrupt_addr, and interrupt_mask.
Prior to cpu stepping, the bits indicated by interrupt_addr and interrupt_mask
are set. Then, after each cpu step, we check to see if the firmware code has
cleared the indicated interrupt bits. If so, we stop executing early, before
N_max_steps is reached, and return the number of executed cycles.

This was done so that we could stop executing cpu cycles whenever the user's
gnc() function returned.